Move MOSS code to `plagcheck` module
